# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Titan-V API
  version: 1.0.0
  description: |
    Backend for the Titan-V geospatial terminal UI: protocols index, geocode proxy,
    weather proxy, target registry, and health endpoints.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/targets:
    get:
      tags: [Targets]
      summary: List coordinate registry entries
      operationId: listTargets
      responses:
        "200":
          description: Targets
          content:
            application/json:
              schema:
                type: object
                required: [targets]
                properties:
                  targets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Target"
    post:
      tags: [Targets]
      summary: Create a registry entry (after geocode on the client)
      operationId: createTarget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTargetRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [target]
                properties:
                  target:
                    $ref: "#/components/schemas/Target"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/targets/{id}:
    delete:
      tags: [Targets]
      summary: Remove a registry entry
      operationId: deleteTarget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    Target:
      type: object
      required: [id, name, lat, lon, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        lat:
          type: number
        lon:
          type: number
        createdAt:
          type: string
          format: date-time
    CreateTargetRequest:
      type: object
      required: [name, lat, lon]
      properties:
        name:
          type: string
        lat:
          type: number
        lon:
          type: number